updating to cpp14 version

+++ src/KOKKOS/atom_vec_demsi_kokkos.cpp
@@ -4037,9 +4037,9 @@ int AtomVecDemsiKokkos::write_vel_hybrid(FILE *fp, double *buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecDemsiKokkos::memory_usage()
+double AtomVecDemsiKokkos::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
